backend/requirements.txt
aiofiles==0.7.0
aiomcache==0.6.0
anyio==3.3.0
asgiref==3.4.1
autopep8==1.5.7
blinker==1.4
bson==0.5.10
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Deprecated==1.2.12
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
priority==2.0.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
Quart==0.15.1
Quart-CORS==0.5.0
Quart-Session==1.0.2
requests==2.26.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
urllib3==1.26.6
uvicorn==0.14.0
Werkzeug==2.0.1
wrapt==1.12.1
wsproto==1.0.0
